Decouple TTY library from terminal
