+++ mosaic-terminal/src/commonMain/kotlin/com/jakewharton/mosaic/terminal/PlatformEventHandler.kt
@@ -4,6 +4,7 @@ import com.jakewharton.mosaic.terminal.event.DebugEvent
 import com.jakewharton.mosaic.terminal.event.Event
 import com.jakewharton.mosaic.terminal.event.FocusEvent
 import com.jakewharton.mosaic.terminal.event.ResizeEvent
+import com.jakewharton.mosaic.tty.Tty
 import kotlinx.coroutines.channels.SendChannel
 
 internal class PlatformEventHandler(
