--- mosaic-terminal/src/jvmMain/resources/META-INF/proguard/com.jakewharton.mosaic-terminal.pro
@@ -1,12 +1,0 @@
-# Note: Our JNI code only creates JDK types and only uses Java built-in types across the boundary.
--keep class com.jakewharton.mosaic.terminal.Jni {
-  native <methods>;
-}
-
-# These members are interacted with through native code.
--keep class com.jakewharton.mosaic.terminal.Tty {
-	void onFocus(...);
-	void onKey(...);
-	void onMouse(...);
-	void onResize(...);
-}
