+++ mosaic-tty/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=mosaic-tty
+POM_NAME=Mosaic TTY
