+++ mosaic-tty/src/jvmMain/include/LICENSE
