+++ mosaic-tty/src/jvmMain/kotlin/com/jakewharton/mosaic/tty/TestTty.kt
@@ -1,0 +1,55 @@
+package com.jakewharton.mosaic.tty
+
+import com.jakewharton.mosaic.tty.Jni.testTtyGetTty
+import com.jakewharton.mosaic.tty.Jni.testTtyInit
+import com.jakewharton.mosaic.tty.Jni.ttyCallbackFree
+import com.jakewharton.mosaic.tty.Jni.ttyCallbackInit
+
+public actual class TestTty(
+	private var testTtyPtr: Long,
+	public actual val tty: Tty,
+) : AutoCloseable {
+	public actual companion object {
+		public actual fun create(callback: Tty.Callback): TestTty {
+			val callbackPtr = ttyCallbackInit(callback)
+			if (callbackPtr != 0L) {
+				val testTtyPtr = testTtyInit(callbackPtr)
+				if (testTtyPtr != 0L) {
+					val ttyPtr = testTtyGetTty(testTtyPtr)
+					val tty = Tty(ttyPtr, callbackPtr)
+					return TestTty(testTtyPtr, tty)
+				}
+				ttyCallbackFree(callbackPtr)
+			}
+			throw OutOfMemoryError()
+		}
+	}
+
+	public actual fun write(buffer: ByteArray) {
+		Jni.testTtyWrite(testTtyPtr, buffer)
+	}
+
+	public actual fun focusEvent(focused: Boolean) {
+		Jni.testTtyFocusEvent(testTtyPtr, focused)
+	}
+
+	public actual fun keyEvent() {
+		Jni.testTtyKeyEvent(testTtyPtr)
+	}
+
+	public actual fun mouseEvent() {
+		Jni.testTtyMouseEvent(testTtyPtr)
+	}
+
+	public actual fun resizeEvent(columns: Int, rows: Int, width: Int, height: Int) {
+		Jni.testTtyResizeEvent(testTtyPtr, columns, rows, width, height)
+	}
+
+	actual override fun close() {
+		if (testTtyPtr != 0L) {
+			tty.close()
+			Jni.testTtyFree(testTtyPtr)
+			testTtyPtr = 0
+		}
+	}
+}
