+++ mosaic-tty/src/jvmTest/resources/META-INF/proguard/com.jakewharton.mosaic-tty-tests.pro
